[Silver III] Title: 주유소, Time: 364 ms, Memory: 35504 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -19,12 +19,12 @@ public class Main {
         }
 
         Stack<Integer> stack = new Stack<>();
-        int ans = 0;
+        long ans = 0;
         int min = cost[0];
         for(int i = 0; i < N - 1; i++) {
             min = Math.min(min, cost[i]);
 
-            ans += min * distance[i];
+            ans += (long) min * distance[i];
         }
         System.out.println(ans);
     }
